move theme/locale switchers from header, update dependencies
