+++ package.json
@@ -65,6 +65,7 @@
   },
   "dependencies": {
     "clsx": "^1.2.1",
+    "i18next": "^23.6.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-i18next": "^13.3.0"
